Add files via upload
The main "response testrig" page has not changed since we last spoke.

Organizational Settings has added a place to change the signature for the draft emails and a place to put in the response for when things are caught as by the threat responder.

Agent Instructions has stayed much the same, with a new spot for the agent department as a "Base Category". This has not been implemented yet, but I'm getting some ideas on how I can get that sorted.

Workflow Rules has been updated to add some text areas to modify the prompts. This includes being able to change how the research router, search function, and draft email prompts will work. This was just to help modify those parts, and give examples of how we can give the users more power.

Also included, is a graph that shows how the workflow can go. This is partially hardcoded, as I have not found a way to get the conditional edges yet, but I have an idea on how I might be able to do it in the near future. The non-conditional edges are reactive to the nodes we have in the workflow. This includes all the edges except from "categorize_email" to the 3 nodes it's connected to ("research_info_search", "draft_email_writer", "threat_responder").

Model is the tab where you can change the model that is being used. Currently, Llama3-8b, Gemma2-9b, and Mixtral-8x7b are the options, but as this is still currently using Groq, any new ones or higher parameter sizes can be added quickly

One problem I was having for a while that is now resolved is that a lot of the changes you could make before required a save and then a restart of the program. This is no longer the case, as I found out the solution, which was already being used in the code and I had not understood how it worked. Luckily this helped me understand how the code works better.
The only thing that requires a restart is changing which model you want to use.